* do not assume that the remote device is bound to the usb driver
* do not assume that the stdout is a pipe in the remote script

on a socket or pseudo-tty a "broken pipe" condition would return
a POLLHUP rather than POLLERR in poll()

* fix the 'brackets' command line hack which was broken; with it
in place, you could override the local ssh and/or the remote perl
commands with:

usbip-ssh [ ssh_wrapper ... ] user@host [ perl_wrapper ... ] ...

where '[' is a string containing any of the [{<( characters, and ']'
is its mirrored variant, e.g. '{{' and '}}' or 'foo<' and '>oof'.

for instance, use:

usbip-ssh user@host [ sudo perl ] 5-1.2.4

on systems where you have to login as a regular user and then use
'sudo' to become root.